refactor: :recycle: Move codes into prompt package
