+++ include/minishell.h
@@ -1,20 +1,16 @@
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
-// for window
 # include <sys/stat.h>
 # include <fcntl.h>
-
-# include <stdio.h>
-# include <stdlib.h>
 # include <sys/types.h>
 # include <sys/dirent.h>
+# include <termios.h>
 
 # include "libft.h"
 
 # include <readline/readline.h>
 # include <readline/history.h>
 
-// # include "get_next_line.h"
-
+# include "prompt.h"
 #endif
