+++ src/prompt/interrupt.c
@@ -1,0 +1,17 @@
+#include "minishell.h"
+
+/*
+	- printf: move to new line
+	- rl_on_new_line: Regenerate the prompt on a newline
+	- rl_replace_line: Clear the previous text
+*/
+void	signal_handler(int status)
+{
+	if (status == SIGINT)
+	{
+		printf("\n");
+		rl_on_new_line();
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}
+}
