+++ src/prompt/terminal.c
@@ -1,0 +1,1 @@
+#include "minishell.h"
